# PERF: Replace `.shape` with `.__len__`