```python
In [4]: timeit len(df.columns)
306 ns ± 5.61 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [5]: timeit df.shape[1]
800 ns ± 5.31 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```

<!--
Thanks for contributing a pull request!

Please follow these standard acronyms to start the commit message:

- ENH: enhancement
- BUG: bug fix
- DOC: documentation
- TYP: type annotations
- TST: addition or modification of tests
- MAINT: maintenance commit (refactoring, typos, etc.)
- BLD: change related to building
- REL: related to releasing
- API: an (incompatible) API change
- DEP: deprecate something, or remove a deprecated object
- DEV: development tool or utility
- REV: revert an earlier commit
- PERF: performance improvement
- BOT: always commit via a bot
- CI: related to CI or CD
- CLN: Code cleanup
-->

- [ ] closes #xxxx
- [x] whatsnew entry